Add logging (#23)
* add basic logging support

This adds a new helper package named `logging` which contains a few
helper functions. One is a singleton wrapper to return a global logging
object, and the other is a decorator to wrap http handlers with a
request logger.

* add a helper for creating uuids

This change adds a small helper package for generating version 4 UUIDs.

* add responses to logs

This change adds a bit to the middleware for the logging which allows
the printing of return status codes for http requests. To help align the
logs, a UUID request identifier has been added to each call.

* add ability to specify a log file

This change adds a command line option for creating, or appending to, a
log file. It will channel all logs created through the oshinko-rest
specific logger into the specified file, creating it if necessary. By
default, and on file error, it will log to stderr.

* add some error detection around command line flags

This change adds a simple test to ensure that we are not attempting to
operate on the wrong command line options group.

* update SetLoggerFile comments

This change makes it clear in the comments that on a failed call to
SetLoggerFile the stderr logger will be set.