restructure current tests
This change moves the current tests to a separate package and customizes
the build scripts with a new testing specific script. The current tests
are moved to a new package because they are specifically client/server
related and may not pertain to the unit tests. The testing script was
added to help better refine the scripts we use for building and testing.
This gives us a more granular approach to our builds and tests.

Changes
* move current tests to the tests/client directory
* change tests package to clienttest package
* remove test related script from tools/build.sh
* add tools/test.sh with test related script
* change Makefile to use new test.sh script
* add tools/common.sh to configure the link flags for both build.sh and
  test.sh