Add a simple bash script to setup an oshinko dev env
This is a bit rough and can certainly use improvement,
but it's a baseline for getting oshinko up and running
on a fresh VM. Tested this on F23.

Note, it's relatively idempotent. Running this multiple
times from the same root directory showed no problems
(except that your local image registry will fill up
with tags :) )

* Clone the oshinko repos
* build all the images
* install an openshift cluster with oc cluster up
* create an "oshinko" user and project
* tag and push all the images
* launch the standard server-ui-template.yaml