Add more unittests (#32)
* restructure current tests

This change moves the current tests to a separate package and customizes
the build scripts with a new testing specific script. The current tests
are moved to a new package because they are specifically client/server
related and may not pertain to the unit tests. The testing script was
added to help better refine the scripts we use for building and testing.
This gives us a more granular approach to our builds and tests.

Changes
* move current tests to the tests/client directory
* change tests package to clienttest package
* remove test related script from tools/build.sh
* add tools/test.sh with test related script
* change Makefile to use new test.sh script
* add tools/common.sh to configure the link flags for both build.sh and
  test.sh

* add uuid unit test

This change add the first unit test, and the unittest package. It also
has some housekeeping for the test runner.

Changes
* add unittest target to tools/test.sh
* clean up syntax in tools/test.sh
* add unit test package with configuration for gocheck
* add test for the uuid package

* add information about unit tests to contributing

This change adds a description and example of how unit tests can be
added to the project. It does not cover the client/server tests as that
infrastructure needs more work.

* add tests for containers helper

This change adds tests for all the functions in the containers package.
It also does a minor name refactor on 2 of the functions in containers
for consistency.

* add unittests for deploymentconfigs package

This add unit tests for all the functions and structures in the
deploymentconfigs package, it also fixes one minor bug in the FindPort
function.

* add placeholder for authentication tests

* add unit tests for info package

* add placeholder for logging tests

Also fix a spelling error in the comments

* add unit tests for podtemplates package

Also fix a whitespace error in the podtemplates file.

* add unit tests for services package

Also add some minor name cleanups to help with keeping the tests
separate.

* fix word inversion in contributing doc