Updating oshinko-setup.sh to use the Spark image (#46)
* Updating oshinko-setup.sh to use a bespoke Spark
image rather than openshift-spark.  As a part of this
the env vars for worker nodes must also include
the SPARK_MASTER_ADDRESS, so I have also made
a tweak to clusters.go to make that happen.

* Updating oshinko-setup.sh to optionally accept an
extra -s parameter that will specify the pre-built
spark docker image to use to form clusters.  If that
option is not given, this script will build the
openshift-spark image and use that instead.

* Adding oshinko-setup.sh to README.md

Adding some mention of oshinko-setup.sh in the
README.md file.  Covering usage of -w and -s flags
along with some other basic notes about the behavior.

* Tweaks to Debugging.md

Trading an unbalanced paren for a comma.