remove empty tests
The original intent of the empty tests was to leave room for mocked
tests. Unfortunately, it appears that the mocking route will not bear
many fruits and as such these files are being removed and the mock
comments changed in the info tests. Also removing the authentication
and logging helpers from the coverage script.