Initial implementation for named cluster configs
This change adds a "config" object to the NewCluster model which
now contains the master and worker counts along with a name field.
If the name is set, oshinko will look for the named configuration at
/etc/oshinko-cluster-configs and use that as the configuration for
the cluster with master and worker counts applied if they are
non-zero. For full details, see the cluster-configs.md spec
in the oshinko-specs repo.