Config object should contain configured master/worker count
The JSON responses from create, update, and find operations
should contain the configured master and worker node count, not
the actualy master and worker node count. The current master
and worker counts can be figured out from the pod list if
it's important. This is particularly important on create and
update where the response from the operation should inform
the caller of how many pods there will be (not how many
currently exist).