add commit hash to version (#86)
* add commit hash to version

This change will add the 7 digit commit hash to the tail of the tag name
when it is computed by the build tools. It also changes the default git
tag name to "unknown", this better reflects situations where the tags
have not be pulled from the source repo.

Changes
* update common script library tagname function to reflect commit hashes

* add a log message about server version

This is another convenience to help with assuring which version of a
server is installed.