Create an empty oshinko-spark-driver-config ConfigMap
This empty ConfigMap will be mounted by default on
spark driver pods by the oshinko s2i templates. The presence
of an empty config directory will cause the start.sh
script to use the default spark configuration in the image.
If this ConfigMap is edited or a different one is used,
start.sh will use the config directory as the spark config.